+++ src/main/kotlin/d2t/terra/abubaria/light/LightRect.kt
@@ -1,3 +1,3 @@
 package d2t.terra.abubaria.light
 
-class LightRect(val x: Int,val y: Int,val width: Int,val height: Int,val power: Int)
+class LightRect(val x: Float,val y: Float,val width: Float,val height: Float,val power: Int)
